requirements.txt
absl-py==0.14.1
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
attrs==21.2.0
billiard==3.3.0.23
cached-property==1.5.2
celery==3.1.25
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
dataclasses==0.6
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap3==15.0.0
django-import-export==2.7.1
djongo==1.3.6
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.1.1
joblib==1.1.0
kiwisolver==1.3.2
kombu==3.0.37
MarkupPy==1.14
matplotlib==3.3.4
mediapipe==0.8.3
mongoengine==0.23.1
mysql==0.0.3
mysqlclient==2.1.0
numpy==1.19.3
odfpy==1.4.1
opencv-python==4.5.1.48
openpyxl==3.0.9
pandas==1.3.5
Pillow==8.1.0
protobuf==3.18.0
psycopg2-binary==2.8.6
pymongo==3.12.3
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
qrcode==7.3.1
requests==2.25.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.2.4
tablib==3.0.0
threadpoolctl==3.0.0
tqdm==4.56.0
urllib3==1.26.7
vine==5.0.0
whitenoise==5.2.0
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
zipp==3.6.0
